test(4-gpu-b200): split test_qwen35_models.py into 3 files + bump partitions to 6
Splits `test_qwen35_models.py` into one file per test class so the
CI auto-partitioner can distribute the load across separate partitions,
and bumps the `stage-c-test-4-gpu-b200` partition count from 5 to 6
because the total est_time for the suite now slightly exceeds
5 * 30-min step cap.

Previously all three test classes (TestQwen35FP4, TestQwen35FP4MTP,
TestQwen35FP4MTPV2) lived in one file registered with est_time=768s.
Each class spins up its own server on the 234 GB Qwen3.5-397B-A17B-NVFP4
model (3 server launches total). Actual wall clock:

  - Faster B200 host: ~21 min total - fits the 30-min step timeout
  - Slower B200 host: ~28-36 min - times out partition (3) of
            stage-c-test-4-gpu-b200 because 3x server loads + 3x gsm8k
            bodies cross the 30-min cap

est_times are calibrated from a passing slower-host run, per class:

  - Triton (no spec decode, gsm8k ~3x slower generation): ~720s
  - MTP (NEXTN spec decode, ~107s gsm8k body):             ~540s
  - MTPV2 (NEXTN + SGLANG_ENABLE_SPEC_V2, ~104s gsm8k):    ~540s

Split layout:
  - test_qwen35_fp4_triton.py   TestQwen35FP4        (est_time=720)
  - test_qwen35_fp4_mtp.py      TestQwen35FP4MTP     (est_time=540)
  - test_qwen35_fp4_mtp_v2.py   TestQwen35FP4MTPV2   (est_time=540)

Partition count bump (5 -> 6): total est_time for
stage-c-test-4-gpu-b200 suite is now ~9187s > 9000s (= 5 * 30min cap),
so 5 partitions would each average 30+ min. With 6 partitions the
average is ~25.5 min, comfortably below the step cap.

No test is dropped or modified - the three servers still run as before,
just not in the same CI job.
